Fix build for platforms with USE_FUNCLET_CALL_HELPER
I've missed the fact that the macro is defined in the
exceptionhandling.cpp only, so the move to eetwain.cpp disabled this
feature.
